Logger moved to tools folder
